% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{bedMethylFields}
\alias{bedMethylFields}
\title{List retrievable bedMethyl fields}
\usage{
bedMethylFields()
}
\value{
A \code{data.frame} with columns \code{field}, \code{type}, and \code{description}.
}
\description{
Returns a data.frame describing retrievable bedMethyl fields and their types.
}
\examples{
bedMethylFields()
}
