% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{filterByCoverage}
\alias{filterByCoverage}
\alias{filterByCoverage,RBedMethyl-method}
\title{Filter by coverage}
\usage{
filterByCoverage(x, min_cov)
}
\arguments{
\item{x}{An \code{RBedMethyl} object.}

\item{min_cov}{Minimum coverage threshold.}
}
\value{
A filtered \code{RBedMethyl} object.
}
\description{
Filter an \code{RBedMethyl} object by minimum coverage.
}
\examples{
lines <- c(
  paste("chr1", 0, 1, "m", 0, "+", 0, 1, 0, 10, 0.5, 5, 5, 0, 0, 0, 0, 0, sep = "\t"),
  paste("chr1", 10, 11, "m", 0, "+", 10, 11, 0, 20, 0.25, 5, 15, 0, 0, 0, 0, 0, sep = "\t")
)
tmp <- tempfile(fileext = ".bed")
writeLines(lines, tmp)
bm <- readBedMethyl(tmp, mod = "m", fields = c("coverage", "pct", "mod_reads"))
bm2 <- filterByCoverage(bm, min_cov = 15)
length(RBedMethyl::beta(bm2))
}
