% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBedMethyl-class.R
\docType{class}
\name{RBedMethyl-class}
\alias{RBedMethyl-class}
\title{RBedMethyl class}
\description{
Disk-backed representation of nanoporetech modkit bedMethyl data from ONT sequencing.
}
\section{Slots}{

\describe{
\item{\code{assays}}{A \code{SimpleList} of assay arrays.}

\item{\code{chrom_levels}}{Character vector of chromosome names.}

\item{\code{strand_levels}}{Character vector of strand levels.}

\item{\code{chr_index}}{Matrix of chromosome row ranges (start/end).}

\item{\code{index}}{Integer vector of active row indices.}

\item{\code{mod}}{Modification code.}
}}

