% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[,RBedMethyl,missing,missing,missing-method}
\alias{[,RBedMethyl,missing,missing,missing-method}
\title{Subset rows}
\usage{
\S4method{[}{RBedMethyl,missing,missing,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An \code{RBedMethyl} object.}

\item{i}{Integer, logical, or \code{GRanges} index.}

\item{j}{Unused.}

\item{...}{Unused.}

\item{drop}{Unused.}
}
\value{
A filtered \code{RBedMethyl} object.
}
\description{
Subset an \code{RBedMethyl} object by integer, logical, or \code{GRanges} index.
}
