% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subsetBy}
\alias{subsetBy}
\alias{subsetBy,RBedMethyl-method}
\title{Subset by assay predicate}
\usage{
subsetBy(x, column, FUN)
}
\arguments{
\item{x}{An \code{RBedMethyl} object.}

\item{column}{Assay name to filter on (must be loaded).}

\item{FUN}{Predicate function returning a logical vector.}
}
\value{
A filtered \code{RBedMethyl} object.
}
\description{
Subset an \code{RBedMethyl} object using a predicate over an assay.
}
\examples{
lines <- c(
  paste("chr1", 0, 1, "m", 0, "+", 0, 1, 0, 10, 0.5, 5, 5, 0, 0, 0, 0, 0, sep = "\t"),
  paste("chr1", 10, 11, "m", 0, "+", 10, 11, 0, 20, 0.25, 5, 15, 0, 0, 0, 0, 0, sep = "\t")
)
tmp <- tempfile(fileext = ".bed")
writeLines(lines, tmp)
bm <- readBedMethyl(tmp, mod = "m", fields = c("coverage", "pct", "mod_reads"))
bm2 <- subsetBy(bm, "coverage", function(v) v >= 15)
length(RBedMethyl::beta(bm2))
}
