% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subsetByChromosomes}
\alias{subsetByChromosomes}
\alias{subsetByChromosomes,RBedMethyl-method}
\title{Subset by chromosomes}
\usage{
subsetByChromosomes(x, chr)
}
\arguments{
\item{x}{An \code{RBedMethyl} object.}

\item{chr}{Character vector of chromosome names.}
}
\value{
A filtered \code{RBedMethyl} object.
}
\description{
Subset an \code{RBedMethyl} object by one or more chromosomes.
}
\examples{
lines <- c(
  paste("chr1", 0, 1, "m", 0, "+", 0, 1, 0, 10, 0.5, 5, 5, 0, 0, 0, 0, 0, sep = "\t"),
  paste("chr2", 10, 11, "m", 0, "+", 10, 11, 0, 20, 0.25, 5, 15, 0, 0, 0, 0, 0, sep = "\t")
)
tmp <- tempfile(fileext = ".bed")
writeLines(lines, tmp)
bm <- readBedMethyl(tmp, mod = "m", fields = c("coverage", "pct", "mod_reads"))
bm2 <- subsetByChromosomes(bm, c("chr1"))
length(RBedMethyl::beta(bm2))
}
