% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subsetByRegion,RBedMethyl,GRanges,missing,missing-method}
\alias{subsetByRegion,RBedMethyl,GRanges,missing,missing-method}
\alias{subsetByRegion,RBedMethyl,GRanges-method}
\title{Subset by GRanges}
\usage{
\S4method{subsetByRegion}{RBedMethyl,GRanges,missing,missing}(x, chr, start, end)
}
\arguments{
\item{x}{An \code{RBedMethyl} object.}

\item{chr}{A \code{GRanges} object of regions.}

\item{start}{Unused (for signature compatibility).}

\item{end}{Unused (for signature compatibility).}
}
\value{
A filtered \code{RBedMethyl} object.
}
\description{
Subset an \code{RBedMethyl} object by overlaps with a \code{GRanges}.
}
