% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subsetByRegion}
\alias{subsetByRegion}
\alias{subsetByRegion,RBedMethyl-method}
\alias{subsetByRegion,RBedMethyl,character,numeric,numeric-method}
\title{Subset by region}
\usage{
subsetByRegion(x, chr, start, end)
}
\arguments{
\item{x}{An \code{RBedMethyl} object.}

\item{chr}{Chromosome name.}

\item{start}{Region start (0-based, half-open).}

\item{end}{Region end.}
}
\value{
A filtered \code{RBedMethyl} object.
}
\description{
Subset an \code{RBedMethyl} object by genomic interval.
}
\examples{
lines <- c(
  paste("chr1", 0, 1, "m", 0, "+", 0, 1, 0, 10, 0.5, 5, 5, 0, 0, 0, 0, 0, sep = "\t"),
  paste("chr1", 10, 11, "m", 0, "+", 10, 11, 0, 20, 0.25, 5, 15, 0, 0, 0, 0, 0, sep = "\t")
)
tmp <- tempfile(fileext = ".bed")
writeLines(lines, tmp)
bm <- readBedMethyl(tmp, mod = "m", fields = c("coverage", "pct", "mod_reads"))
bm2 <- subsetByRegion(bm, "chr1", 0, 5)
length(RBedMethyl::beta(bm2))
}
