% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summarizeByRegion}
\alias{summarizeByRegion}
\alias{summarizeByRegion,RBedMethyl-method}
\title{Summarize by regions}
\usage{
summarizeByRegion(x, regions)
}
\arguments{
\item{x}{An \code{RBedMethyl} object.}

\item{regions}{A \code{GRanges} of regions.}
}
\value{
A \code{DataFrame} with coverage, mod_reads, beta, and n_sites.
}
\description{
Summarize methylation by a set of regions.
}
\examples{
lines <- c(
  paste("chr1", 0, 1, "m", 0, "+", 0, 1, 0, 10, 0.5, 5, 5, 0, 0, 0, 0, 0, sep = "\t"),
  paste("chr1", 10, 11, "m", 0, "+", 10, 11, 0, 20, 0.25, 5, 15, 0, 0, 0, 0, 0, sep = "\t")
)
tmp <- tempfile(fileext = ".bed")
writeLines(lines, tmp)
h5 <- tempfile(fileext = ".h5")
bm <- readBedMethyl(tmp, h5, mod = "m", fields = c("coverage", "pct", "mod_reads"))
regions <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(start = 1, end = 12)
)
summarizeByRegion(bm, regions)
}
